Done with extraction

+++ db/migrate/005_add_mobile_phone_to_users.rb
@@ -1,13 +1,9 @@
 class AddMobilePhoneToUsers < ActiveRecord::Migration
   def up
-    unless column_exists? :users, :mobile_phone
-      add_column :users, :mobile_phone, :string
-    end
+    add_column :users, :mobile_phone, :string unless column_exists? :users, :mobile_phone
   end
 
   def down
-    unless Redmine::Plugin.installed?(:redmine_sms_auth)
-      remove_column :users, :mobile_phone
-    end
+    remove_column :users, :mobile_phone unless Redmine::Plugin.installed?(:redmine_sms_auth)
   end
 end
